+++ Qtica/enums/__init__.py
@@ -11,3 +11,4 @@ from .smooth_scroll import SmoothMode
 from .teaching_tip_tails import TeachingTipTailPositions
 from .clipboard import ClipboardDataTypes
 from .theme import Theme
+from ._abs_icons import AbstractIcons