+++ Qtica/tools/icon.py
@@ -1,20 +1,25 @@
 from typing import Union
 from PySide6.QtCore import QFileInfo, QSize, Qt
-from PySide6.QtGui import QIcon, QIconEngine, QImage, QPixmap, QPainter
+from PySide6.QtGui import QIcon, QIconEngine, QImage, QPixmap, QPainter, QColor
 from PySide6.QtWidgets import QFileIconProvider
 from ..core.base import BehaviorDeclarative
-from .color import Color
+from ..enums._abs_icons import AbstractIcons
 
 
 class Icon(BehaviorDeclarative):
     def __init__(self, 
-                 icon: Union[QIcon, 
-                             QIconEngine, 
-                             str, 
-                             QPixmap],
-                 color: Color = None,
+                 icon: Union[str,
+                             QIcon,
+                             QIconEngine,
+                             QPixmap,
+                             QImage,
+                             AbstractIcons],
+                 color: QColor = None,
                  size: QSize = None,
                  **kwargs) -> QIcon:
+
+        if isinstance(icon, AbstractIcons):
+            icon = icon.value
 
         if (isinstance(icon, str)
             and icon.endswith(".svg")
@@ -26,19 +31,34 @@ class Icon(BehaviorDeclarative):
 
         return QIcon(icon)
 
-    def _colored_icon(self, 
-                      icon, 
-                      color: Color = None) -> QIcon:
+    def _colored_icon(self,
+                      icon: Union[QPixmap, QImage, str],
+                      color: QColor = None,
+                      size: QSize = None) -> QIcon:
 
-        pixmap = QPixmap(icon)
+        if isinstance(icon, (QIcon, QIconEngine)):
+            pixmap = icon.pixmap(size
+                                if size is not None 
+                                else icon.availableSizes()[0])
+        else:
+            pixmap = QPixmap(icon)
+            if size is not None:
+                pixmap = pixmap.scaled(size, 
+                                       Qt.AspectRatioMode.KeepAspectRatio)
+
         pixmap_painter = QPainter(pixmap)
         pixmap_painter.setCompositionMode(QPainter.CompositionMode.CompositionMode_SourceIn)
-        pixmap_painter.fillRect(pixmap.rect(), color if color is not None else -1)
+        pixmap_painter.fillRect(pixmap.rect(),
+                                color 
+                                if color is not None 
+                                else -1)
         pixmap_painter.end()
         return QIcon(pixmap)
 
     @staticmethod
-    def make_svg(svg: str, color: Color = None) -> QIcon:
+    def make_svg(svg: str, 
+                 color: QColor = None) -> QIcon:
+
         image = QImage()
         image.fill(Qt.GlobalColor.transparent)
         image.loadFromData(bytes(svg, "utf-8"))